refactored RTS for jvm into CommonRTS
